// src/api/types/CreateEvaluatorLogResponse.ts
/**
 * This file was auto-generated by Fern from our API Definition.
 */

export interface CreateEvaluatorLogResponse {
    /** String identifier of the new Log. */
    id: string;
    /** Identifier of the evaluated parent Log. */
    parentId: string;
    /** Identifier of the Session containing both the parent and the new child Log. If the parent Log does not belong to a Session, a new Session is created with this ID. */
    sessionId?: string;
    /** Identifier of Evaluator Version for which the Log was registered. */
    versionId: string;
}