// src/api/types/CreatePromptLogResponse.ts
/**
 * This file was auto-generated by Fern from our API Definition.
 */

export interface CreatePromptLogResponse {
    /** String ID of log. */
    id: string;
    /** ID of the Prompt the log belongs to. */
    promptId: string;
    /** ID of the specific version of the Prompt. */
    versionId: string;
    /** String ID of session the log belongs to. */
    sessionId?: string;
}